requirements.txt
discord.py>=0.16.0,<0.17.0
